added -Wall, fixed bugs, moving scrubber
git-svn-id: https://svn.csail.mit.edu/drc@1383 c7283977-0100-402a-a91a-fa70b306dbfe

+++ software/action_authoring/src/qt4_widgets/CustomSlider.h
@@ -26,7 +26,7 @@ class DefaultValueSlider : public QSlider {
   int getSliderPosition(int index) {
       if (index < 0)
 	  return 0;
-      if (index >= _tick_positions.size())
+      if (index >= (int)_tick_positions.size())
 	  return width();
 
       return QStyle::sliderPositionFromValue(
